requirements.txt
requests==2.28.2
python-binance==1.0.17
